refactor(stack): establish explicit domain model
