+++ packages/stack/src/ServiceName.ts
@@ -1,0 +1,15 @@
+/** Public Supabase services represented in a local stack. */
+export type ServiceName =
+  | "postgres"
+  | "postgrest"
+  | "auth"
+  | "edge-runtime"
+  | "realtime"
+  | "storage"
+  | "imgproxy"
+  | "mailpit"
+  | "pgmeta"
+  | "studio"
+  | "analytics"
+  | "vector"
+  | "pooler";
