+++ packages/stack/src/createStack.unit.test.ts
@@ -2,12 +2,18 @@ import { describe, expect, it } from "vitest";
 import { mkdtempSync, mkdirSync, rmSync, writeFileSync } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
-import type { ReadyOptions, StackHandle } from "./createStack.ts";
-import { resolveConfig, resolveDaemonConfig } from "./createStack.ts";
+import type { StackHandle } from "./createStack.ts";
 import type { AllocatedPorts } from "./PortAllocator.ts";
 import { DEFAULT_MANAGED_STACK_NAME, projectKeyForProjectDir } from "./paths.ts";
 import { stackMetadata } from "./StackMetadata.ts";
-import type { AuthConfig, PostgresConfig, PostgrestConfig, StackConfig } from "./StackBuilder.ts";
+import type {
+  AuthConfig,
+  PostgresConfig,
+  PostgrestConfig,
+  ReadyOptions,
+  StackConfig,
+} from "./StackConfig.ts";
+import { resolveConfig, resolveDaemonConfig } from "./StackConfigResolver.ts";
 import { DEFAULT_VERSIONS } from "./versions.ts";
 
 const DEFAULT_PORTS: AllocatedPorts = {
@@ -249,5 +255,17 @@ describe("resolveConfig startup mode", () => {
   it("preserves an explicit lazy startup mode", async () => {
     const config = await resolveConfig({ startupMode: "lazy" });
     expect(config.startupMode).toBe("lazy");
+  });
+});
+
+describe("resolveConfig readiness policy", () => {
+  it("uses a finite package default", async () => {
+    const config = await resolveConfig();
+    expect(config.readiness).toEqual({ mode: "finite", timeoutMs: 120_000 });
+  });
+
+  it("preserves an explicit infinite policy", async () => {
+    const config = await resolveConfig({ readiness: { mode: "infinite" } });
+    expect(config.readiness).toEqual({ mode: "infinite" });
   });
 });
