+++ packages/stack/src/node.ts
@@ -19,7 +19,7 @@ import {
 } from "./prefetch.ts";
 import { defaultCacheRoot } from "./paths.ts";
 import { StackPreparation } from "./StackPreparation.ts";
-import type { StackConfig } from "./StackBuilder.ts";
+import type { StackConfig } from "./StackConfig.ts";
 import { UnixHttpClient, UnixHttpClientError } from "./UnixHttpClient.ts";
 
 const mergeBodyHeaders = (
